Support char type in HDF5 engine. Data is declared as H5T_NATIVE_UCHAR but is stored in HDF5 files as H5T_NATIVE_U8LE which is still platform dependent.

+++ source/adios2/toolkit/interop/hdf5/HDF5Common.h
@@ -49,7 +49,7 @@ public:
             return;
         }
 
-        for (int i = 0; i < m_Chain.size() - 1; i++)
+        for (size_t i = 0; i < m_Chain.size() - 1; i++)
         {
             H5Gclose(m_Chain[i]);
         }
